refactor: change directory for all project, make different subprojects (api and plugin)
